Move: clock-only related types and traits to clock/types.rs
