+++ hal/src/thumbv7em/clock/types.rs
@@ -1,7 +1,59 @@
 use core::marker::PhantomData;
-use typenum::U0;
+use core::ops::{Add, Sub};
+use typenum::{Add1, Sub1, Unsigned, B1, U0};
 
-use crate::typelevel::{Counter, Decrement, Increment, PrivateDecrement, PrivateIncrement, Sealed};
+use crate::typelevel::Sealed;
+
+mod private {
+    use super::*;
+    pub trait Increment: Counter {
+        type Inc: Counter;
+        fn inc(self) -> Self::Inc;
+    }
+    pub trait Decrement: Counter {
+        type Dec: Counter;
+        fn dec(self) -> Self::Dec;
+    }
+}
+
+pub(crate) use private::Decrement as PrivateDecrement;
+pub(crate) use private::Increment as PrivateIncrement;
+
+/// TODO
+pub trait Increment: PrivateIncrement {}
+
+/// TODO
+pub trait Decrement: PrivateDecrement {}
+
+/// TODO
+impl<N> PrivateIncrement for N
+where
+    N: Sealed + Unsigned + Add<B1>,
+    Add1<N>: Sealed + Unsigned,
+{
+    type Inc = Add1<N>;
+    fn inc(self) -> Self::Inc {
+        Self::Inc::default()
+    }
+}
+
+/// TODO
+impl<N> PrivateDecrement for N
+where
+    N: Sealed + Unsigned + Sub<B1>,
+    Sub1<N>: Sealed + Unsigned,
+{
+    type Dec = Sub1<N>;
+    fn dec(self) -> Self::Dec {
+        Self::Dec::default()
+    }
+}
+
+/// TODO
+pub trait Counter: Sealed {}
+
+/// TODO
+impl<N: Unsigned + Sealed> Counter for N {}
 
 pub struct Counted<T, N: Counter>(pub(crate) T, PhantomData<N>);
