+++ hal/src/thumbv7em/clock/v2/pclk.rs
@@ -9,11 +9,11 @@ use crate::pac;
 
 pub use crate::pac::gclk::pchctrl::GEN_A as PclkSourceEnum;
 
+use crate::clock::types::{Counted, Counter, Decrement, Increment};
 use crate::clock::v2::{Source, SourceMarker};
 use crate::sercom::v2::*;
 use crate::time::Hertz;
-use crate::typelevel::counted::Counted;
-use crate::typelevel::{Counter, Decrement, Increment, Sealed};
+use crate::typelevel::Sealed;
 
 use super::gclk::*;
 use super::sources::dpll::{Pll0, Pll1};
